# bpo-46848: Use stringlib/fastsearch in mmap