--- a/Include/internal/pycore_bytesobject.h
+++ b/Include/internal/pycore_bytesobject.h
@@ -15,12 +15,12 @@ extern PyStatus _PyBytes_InitTypes(PyInterpreterState *);
 
 /* Substring Search */
 
-PyAPI_FUNC(Py_ssize_t)
+extern Py_ssize_t
 _PyBytes_Find(const char *haystack, Py_ssize_t len_haystack,
               const char *needle, Py_ssize_t len_needle,
               Py_ssize_t offset);
 
-PyAPI_FUNC(Py_ssize_t)
+extern Py_ssize_t
 _PyBytes_ReverseFind(const char *haystack, Py_ssize_t len_haystack,
                      const char *needle, Py_ssize_t len_needle,
                      Py_ssize_t offset);