--- a/Include/internal/pycore_bytesobject.h
+++ b/Include/internal/pycore_bytesobject.h
@@ -13,17 +13,6 @@ extern "C" {
 
 extern PyStatus _PyBytes_InitTypes(PyInterpreterState *);
 
-/* Substring Search */
-
-extern Py_ssize_t
-_PyBytes_Find(const char *haystack, Py_ssize_t len_haystack,
-              const char *needle, Py_ssize_t len_needle,
-              Py_ssize_t offset);
-
-extern Py_ssize_t
-_PyBytes_ReverseFind(const char *haystack, Py_ssize_t len_haystack,
-                     const char *needle, Py_ssize_t len_needle,
-                     Py_ssize_t offset);
 
 #ifdef __cplusplus
 }