--- a/Modules/mmapmodule.c
+++ b/Modules/mmapmodule.c
@@ -25,7 +25,6 @@
 #define PY_SSIZE_T_CLEAN
 #include <Python.h>
 #include "pycore_fileutils.h"     // _Py_stat_struct
-#include "pycore_bytesobject.h"   // _PyBytes_Fast(Reverse)Search
 #include "structmember.h"         // PyMemberDef
 #include <stddef.h>               // offsetof()