@rumpelsepp 2022-03-01
I like this proposal better than mine: #31554.