@sweeneyde 2022-03-02
Just a couple of benchmarks on Windows:

```python
from pyperf import Runner
runner = Runner()

for haystack, needle in [
    ("""b'x' * 100_000""", """b'y'"""),
    ("""b'x' * 100_000""", """b'yz'"""),
    ("""b'x' * 100_000""", """b'xy'"""),
    ("""b'x' * 100_000""", """b'yx'"""),
    ("""b'ab' * 100_000""", """b'abracadabra'"""),
    ("""b'a' * 10_000 + b'b' * 10_000 + b'a' * 10_000""",
     """b'a' * 10_001"""),
]:
    runner.timeit(
        f"{needle} in {haystack}",
        setup=f"""\
haystack = {haystack}
needle = {needle}
import mmap
m = mmap.mmap(-1, len(haystack))
m.write(haystack)
m.seek(0)
""",
        stmt=f"m.find(needle)"
    )
```

```
Slower (2):
- b'yx' in b'x' * 100_000: 101 us +- 2 us -> 141 us +- 1 us: 1.39x slower
- b'xy' in b'x' * 100_000: 126 us +- 1 us -> 134 us +- 1 us: 1.06x slower

Faster (4):
- b'a' * 10_001 in b'a' * 10_000 + b'b' * 10_000 + b'a' * 10_000: 15.9 ms +- 0.5 ms -> 71.2 us +- 0.7 us: 223.53x faster
- b'y' in b'x' * 100_000: 101 us +- 1 us -> 3.37 us +- 0.07 us: 30.02x faster
- b'yz' in b'x' * 100_000: 101 us +- 1 us -> 25.5 us +- 0.3 us: 3.95x faster
- b'abracadabra' in b'ab' * 100_000: 335 us +- 8 us -> 303 us +- 3 us: 1.11x faster

Geometric mean: 5.20x faster
```