@sweeneyde 2022-03-02
No refleaks
```
0:00:00 [1/1] test_mmap
beginning 6 repetitions
123456
......

== Tests result: SUCCESS ==

1 test OK.
```